+++ examples/07_router/src/components/NestedBazPage.tsx
@@ -1,5 +1,6 @@
 const Baz = () => (
   <div>
+    <h2>Nested</h2>
     <h3>Baz</h3>
   </div>
 );
