+++ examples/07_router/src/components/NestedQuxPage.tsx
@@ -1,5 +1,6 @@
 const Qux = () => (
   <div>
+    <h2>Nested</h2>
     <h3>Qux</h3>
   </div>
 );
